requirements.txt
openpyxl==3.1.5
colorama==0.4.6
Django==6.0.4
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
pillow==12.2.0
PyJWT==2.12.1
python-dotenv==1.2.2
qrcode==8.2
sqlparse==0.5.5
tzdata==2026.2
